requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

aiomysql==0.2.0
cryptography==43.0.1
google-genai>=0.1.2

pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
